# IDGS901_API_Balones/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "BalonEvo",
    "version": "v1"
  },
  "paths": {
    "/api/Auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Login"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Auth/usuario": {
      "get": {
        "tags": [
          "Auth"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "usuario",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/api/Auth/verificar": {
      "get": {
        "tags": [
          "Auth"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "usuario",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
...
    "/api/Direccion/{usuario}": {
      "post": {
        "tags": [
          "Direccion"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "usuario",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Direccion"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Direccion"
            }
          }
        }
      },
      "get": {
        "tags": [
          "Direccion"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "usuario",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Clientes": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "nombre": {
          "type": "string"
        },
        "usuario": {
          "type": "string"
        },
        "correo": {
          "type": "string"
        },
        "contrasenia": {
          "type": "string"
        },
        "rol": {
          "type": "string"
        },
        "estatus": {
          "type": "string"
        }
      }
    },
...
    "Direccion": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "nombreCompleto": {
          "type": "string"
        },
        "calleNumero": {
          "type": "string"
        },
        "codigoPostal": {
          "type": "string"
        },
        "telefono": {
          "type": "string"
        },
        "usuario": {
          "$ref": "#/definitions/Clientes"
        }
      }
    },
    "Login": {
      "type": "object",
      "properties": {
        "usuario": {
          "type": "string"
        },
        "contrasenia": {
          "type": "string"
        }
      }
    },
...
  }
}